refactor organize hand menu
